requirements/common.txt
PyQt6
PyQt6-tools
sqlmodel
python-decouple
